requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-engineio==3.13.0
python-socketio==4.6.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
websocket-client==0.57.0

pip~=21.1.2
wheel~=0.36.2
Jinja2~=3.1.2
setuptools~=57.0.0
pandas~=1.4.3
SQLAlchemy~=1.4.40
Flask~=2.2.2
numpy~=1.23.1
wget~=3.2
Pillow~=9.5.0
telegram~=0.0.1
telepot~=12.7
PyGithub~=1.58.1
